// cypress/support/proto/asr.d.ts
export interface IVariables {

    /** Variables variables */
    variables?: ({ [k: string]: string }|null);
}
// ...
export interface IUndecodedSeconds {

    /** UndecodedSeconds undecodedSeconds */
    undecodedSeconds?: (number|null);
}
// ...
export interface IFullyFinalized {
}
// ...
export interface IEmotionResult {

    /** EmotionResult name */
    name?: (string|null);

    /** EmotionResult confidence */
    confidence?: (number|null);
}
// ...
export interface IHypothesis {

    /** Hypothesis words */
    words?: (string|null);

    /** Hypothesis acousticCost */
    acousticCost?: (number|null);

    /** Hypothesis linguisticCost */
    linguisticCost?: (number|null);

    /** Hypothesis finalCost */
    finalCost?: (number|null);

    /** Hypothesis phraseStart */
    phraseStart?: (number|null);

    /** Hypothesis phraseEnd */
    phraseEnd?: (number|null);

    /** Hypothesis normalizedText */
    normalizedText?: (string|null);
}
// ...
export interface IDecoderResult {

    /** DecoderResult hypothesis */
    hypothesis?: (IHypothesis[]|null);

    /** DecoderResult chunkStart */
    chunkStart?: (number|null);

    /** DecoderResult chunkEnd */
    chunkEnd?: (number|null);

    /** DecoderResult timeEndpointDetectionMs */
    timeEndpointDetectionMs?: (number|null);

    /** DecoderResult timeDecodingMs */
    timeDecodingMs?: (number|null);

    /** DecoderResult variables */
    variables?: (IVariables|null);

    /** DecoderResult isFinal */
    isFinal?: (boolean|null);

    /** DecoderResult emotionResult */
    emotionResult?: (IEmotionResult[]|null);

    /** DecoderResult contextAnswer */
    contextAnswer?: (DecoderResult.IContextAnswer[]|null);
}

/** Represents a DecoderResult. */
export class DecoderResult implements IDecoderResult {

    /**
     * Constructs a new DecoderResult.
     * @param [properties] Properties to set
     */
    constructor(properties?: IDecoderResult);

    /** DecoderResult hypothesis. */
    public hypothesis: IHypothesis[];

    /** DecoderResult chunkStart. */
    public chunkStart: number;

    /** DecoderResult chunkEnd. */
    public chunkEnd: number;

    /** DecoderResult timeEndpointDetectionMs. */
    public timeEndpointDetectionMs: number;

    /** DecoderResult timeDecodingMs. */
    public timeDecodingMs: number;

    /** DecoderResult variables. */
    public variables?: (IVariables|null);

    /** DecoderResult isFinal. */
    public isFinal: boolean;

    /** DecoderResult emotionResult. */
    public emotionResult: IEmotionResult[];

    /** DecoderResult contextAnswer. */
    public contextAnswer: DecoderResult.IContextAnswer[];

    /**
     * Encodes the specified DecoderResult message. Does not implicitly {@link DecoderResult.verify|verify} messages.
     * @param message DecoderResult message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: IDecoderResult, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a DecoderResult message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns DecoderResult
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): DecoderResult;
}

export namespace DecoderResult {

    /** Properties of a ContextAnswer. */
    interface IContextAnswer {

        /** ContextAnswer contextResult */
        contextResult?: (DecoderResult.ContextAnswer.IContextRef[]|null);
    }

    /** Represents a ContextAnswer. */
    class ContextAnswer implements IContextAnswer {

        /**
         * Constructs a new ContextAnswer.
         * @param [properties] Properties to set
         */
        constructor(properties?: DecoderResult.IContextAnswer);

        /** ContextAnswer contextResult. */
        public contextResult: DecoderResult.ContextAnswer.IContextRef[];

        /**
         * Encodes the specified ContextAnswer message. Does not implicitly {@link DecoderResult.ContextAnswer.verify|verify} messages.
         * @param message ContextAnswer message or plain object to encode
         * @param [writer] Writer to encode to
         * @returns Writer
         */
        public static encode(message: DecoderResult.IContextAnswer, writer?: $protobuf.Writer): $protobuf.Writer;

        /**
         * Decodes a ContextAnswer message from the specified reader or buffer.
         * @param reader Reader or buffer to decode from
         * @param [length] Message length if known beforehand
         * @returns ContextAnswer
         * @throws {Error} If the payload is not a reader or valid buffer
         * @throws {$protobuf.util.ProtocolError} If required fields are missing
         */
        public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): DecoderResult.ContextAnswer;
    }

    namespace ContextAnswer {

        /** Properties of a ContextRef. */
        interface IContextRef {

            /** ContextRef id */
            id?: (string|null);

            /** ContextRef index */
            index?: (number|null);

            /** ContextRef originalValue */
            originalValue?: (string|null);

            /** ContextRef predictedValue */
            predictedValue?: (string|null);

            /** ContextRef score */
            score?: (number|null);
        }

        /** Represents a ContextRef. */
        class ContextRef implements IContextRef {

            /**
             * Constructs a new ContextRef.
             * @param [properties] Properties to set
             */
            constructor(properties?: DecoderResult.ContextAnswer.IContextRef);

            /** ContextRef id. */
            public id: string;

            /** ContextRef index. */
            public index: number;

            /** ContextRef originalValue. */
            public originalValue: string;

            /** ContextRef predictedValue. */
            public predictedValue: string;

            /** ContextRef score. */
            public score: number;

            /**
             * Encodes the specified ContextRef message. Does not implicitly {@link DecoderResult.ContextAnswer.ContextRef.verify|verify} messages.
             * @param message ContextRef message or plain object to encode
             * @param [writer] Writer to encode to
             * @returns Writer
             */
            public static encode(message: DecoderResult.ContextAnswer.IContextRef, writer?: $protobuf.Writer): $protobuf.Writer;

            /**
             * Decodes a ContextRef message from the specified reader or buffer.
             * @param reader Reader or buffer to decode from
             * @param [length] Message length if known beforehand
             * @returns ContextRef
             * @throws {Error} If the payload is not a reader or valid buffer
             * @throws {$protobuf.util.ProtocolError} If required fields are missing
             */
            public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): DecoderResult.ContextAnswer.ContextRef;
        }
    }
}

/** Properties of an ErrorResponse. */
export interface IErrorResponse {

    /** ErrorResponse errorMessage */
    errorMessage?: (string|null);
}
// ...
export interface IPacketWrapperFromServer {

    /** PacketWrapperFromServer undecodedSecondsField */
    undecodedSecondsField?: (IUndecodedSeconds|null);

    /** PacketWrapperFromServer fullyFinalizedField */
    fullyFinalizedField?: (IFullyFinalized|null);

    /** PacketWrapperFromServer decoderResultField */
    decoderResultField?: (IDecoderResult|null);

    /** PacketWrapperFromServer errorResponse */
    errorResponse?: (IErrorResponse|null);
}

/** Represents a PacketWrapperFromServer. */
export class PacketWrapperFromServer implements IPacketWrapperFromServer {

    /**
     * Constructs a new PacketWrapperFromServer.
     * @param [properties] Properties to set
     */
    constructor(properties?: IPacketWrapperFromServer);

    /** PacketWrapperFromServer undecodedSecondsField. */
    public undecodedSecondsField?: (IUndecodedSeconds|null);

    /** PacketWrapperFromServer fullyFinalizedField. */
    public fullyFinalizedField?: (IFullyFinalized|null);

    /** PacketWrapperFromServer decoderResultField. */
    public decoderResultField?: (IDecoderResult|null);

    /** PacketWrapperFromServer errorResponse. */
    public errorResponse?: (IErrorResponse|null);

    /** PacketWrapperFromServer MessageType. */
    public MessageType?: ("undecodedSecondsField"|"fullyFinalizedField"|"decoderResultField"|"errorResponse");

    /**
     * Encodes the specified PacketWrapperFromServer message. Does not implicitly {@link PacketWrapperFromServer.verify|verify} messages.
     * @param message PacketWrapperFromServer message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: IPacketWrapperFromServer, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a PacketWrapperFromServer message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns PacketWrapperFromServer
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): PacketWrapperFromServer;
}